ltx_send_message() retry write properly on EAGAIN
When we are doing file transfer we can easily saturate the serial
connection or pipe we write into and get EAGAIN.

The original code just ignored all write errors, which caused lost bytes
and broken communication with kirk.

This commit fixes two bugs in the retry loop:

- properly adjust the buffer position on subsequent write attempts
- do a short sleep on EAGAIN and retry later on

We also print message into stderr when write() failed for any other
reason than EAGAIN so that there is at least some hint of a failure in
that case.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>